dessy comment on moving UserExists
